Group_5/DDMS/backend/computationalUnit/requirements.txt
matplotlib==3.5.2
numpy==1.21.5
opencv_python==4.7.0.72
pandas
pdf2image==1.16.3
# Pillow==9.0.1
Pillow==9.5.0
PyPDF2==3.0.1
pytesseract
tabula_py
